server/requirements.txt
requests
python-dotenv
pymongo
litestar[standard]